requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
opencc==1.1.1
opencc_python_reimplemented==0.1.6
torch==1.11.0+cpu
transformers==4.17.0
beautifulsoup4==4.11.1
discord.py==1.7.3
pandas==1.4.3
python-dotenv==0.20.0
requests==2.28.0
